+++ src/model/person/client/client.h
@@ -18,18 +18,18 @@ public:
     bool isPremium() const;
     float getMeanEvaluation() const;
     unsigned getPoints() const;
-    std::vector<float> getEvaluations() const;
+    std::vector<int> getEvaluations() const;
     void setPremium(bool premium);
     void addPoints(unsigned points);
     void removePoints(unsigned points);
     void resetPoints();
-    void addEvaluation(float evaluation);
+    void addEvaluation(int evaluation);
 
     void print(std::ostream& os, bool showData = true);
 private:
     bool _premium;
     unsigned _points;
-    std::vector<float> _evaluations;
+    std::vector<int> _evaluations;
 };
